+++ pytsite/pip/__init__.py
@@ -9,9 +9,19 @@ __license__ = 'MIT'
 from . import _error as error
 from ._api import get_installed_info, get_installed_version, is_installed, install, uninstall
 
-# Console commands
-from pytsite import lang as _lang, console as _console
-from . import _console_commands
-_lang.register_package(__name__)
-_console.register_command(_console_commands.Install())
-_console.register_command(_console_commands.Uninstall())
+
+def _init():
+    from pytsite import lang, console, update
+    from . import _cc, _eh
+
+    # Resource packages
+    lang.register_package(__name__)
+
+    # Console commands
+    console.register_command(_cc.Install())
+    console.register_command(_cc.Uninstall())
+
+    # Events handlers
+    update.on_update_stage_2(_eh.update_stage_2)
+
+_init()